Link MLP, replace functions
