+++ test.py
@@ -1,2 +1,6 @@
 def same(a):
-	return a
+	a[0] = 1
+
+a = [2, 2, 3]
+same(a)
+print(a)
